make valid module and import build
